+++ test/unit/backend.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+cd ../../src/backend/
+go get backend
+go build -v .
+
+go test -v .